+++ src/main/java/chess/domain/position/AlphaColumn.java
@@ -1,0 +1,4 @@
+package chess.domain.position;
+
+public class AlphaColumn {
+}
