+++ src/main/java/chess/domain/position/NumberRow.java
@@ -1,0 +1,5 @@
+package chess.domain.position;
+
+public class NumberRow {
+
+}
