+++ src/main/java/chess/view/OutputView.java
@@ -1,6 +1,6 @@
 package chess.view;
 
-import chess.domain.Position;
+import chess.domain.position.Position;
 import chess.domain.Result;
 import chess.domain.piece.Piece;
 import java.util.List;
