+++ src/test/java/chess/domain/ChessBoardTest.java
@@ -1,13 +1,11 @@
-package domain;
+package chess.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 
-import chess.domain.ChessBoard;
-import chess.domain.Position;
-import chess.domain.TeamColor;
+import chess.domain.position.Position;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
 import chess.domain.piece.Piece;
