+++ src/test/java/chess/domain/PiecesTest.java
@@ -1,9 +1,8 @@
-package domain;
+package chess.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.Position;
-import chess.domain.TeamColor;
+import chess.domain.position.Position;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
