+++ src/test/java/chess/domain/piece/BlankTest.java
@@ -1,11 +1,9 @@
-package domain.piece;
+package chess.domain.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.domain.Position;
-import chess.domain.piece.Blank;
-import chess.domain.piece.Piece;
+import chess.domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
