+++ src/test/java/chess/domain/position/AlphaColumnTest.java
@@ -1,0 +1,7 @@
+package chess.domain.position;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class AlphaColumnTest {
+
+}
