+++ src/test/java/chess/domain/position/NumberRowTest.java
@@ -1,0 +1,7 @@
+package chess.domain.position;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class NumberRowTest {
+
+}
